chore: splitting out report page tabs into separate component
